% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_treepath.R
\name{embed_treepath}
\alias{embed_treepath}
\title{Obtain the TreePath embedding of observations}
\usage{
embed_treepath(model, data, depth = 8)
}
\arguments{
\item{model}{A unified model from the \code{treeshap} package.}

\item{data}{A \code{data.frame} with the same features as the model.}

\item{depth}{The maximum depth of the TreePath embedding.}
}
\value{
A \code{data.frame} with the TreePath embeddings of the observations.
}
\description{
This function embeds the observations of a dataset into the TreePath space.
}
\examples{
\donttest{
library(xgboost)
data <- fifa20$data[colnames(fifa20$data) != "work_rate"]
target <- fifa20$target

# calculating simple SHAP values
param <- list(objective = "reg:squarederror", max_depth = 3)
xgb_model <- xgboost::xgboost(as.matrix(data),
    params = param, label = target,
    nrounds = 20, verbose = FALSE
)
unified_model <- xgboost.unify(xgb_model, as.matrix(data))
embeddings <- embed_treepath(unified_model, data)
}

}
