% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_treeshap.R
\name{embed_treeshap}
\alias{embed_treeshap}
\title{Obtain the TreeSHAP embedding of observations}
\usage{
embed_treeshap(model, data)
}
\arguments{
\item{model}{A unified model from the \code{treeshap} package.}

\item{data}{A \code{data.frame} with the same features as the model.}
}
\value{
A `data.frame`` with the TreePath embeddings of the observations.
}
\description{
This function embeds the observations of a dataset into the TreeSHAP space.
}
\examples{
\donttest{
library(xgboost)
data <- fifa20$data[colnames(fifa20$data) != "work_rate"]
target <- fifa20$target

# calculating simple SHAP values
param <- list(objective = "reg:squarederror", max_depth = 3)
xgb_model <- xgboost::xgboost(as.matrix(data),
    params = param, label = target,
    nrounds = 20, verbose = FALSE
)
unified_model <- xgboost.unify(xgb_model, as.matrix(data))
embeddings <- embed_treeshap(unified_model, data)
}

}
