% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeshap.R
\name{is.treeshap}
\alias{is.treeshap}
\title{Check whether object is a valid treeshap object}
\usage{
is.treeshap(x)
}
\arguments{
\item{x}{an object to check}
}
\value{
boolean
}
\description{
Does not check correctness of result, only basic checks
}
