% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_unified.R
\name{model_unified_multioutput.object}
\alias{model_unified_multioutput.object}
\title{Unified model representations for multi-output model}
\value{
List consisting of \code{model_unified} objects, one for each individual output of a model. For survival models, the list is named using the time points, for which predictions are calculated.
}
\description{
\code{model_unified_multioutput} object produced by \code{*.unify} or \code{unify} function.
}
\seealso{
\code{\link{unify}}
}
