% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_unified.R
\name{print.model_unified}
\alias{print.model_unified}
\title{Prints model_unified objects}
\usage{
\method{print}{model_unified}(x, ...)
}
\arguments{
\item{x}{a model_unified object}

\item{...}{other arguments}
}
\description{
Prints model_unified objects
}
