% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_unified.R
\name{print.model_unified_multioutput}
\alias{print.model_unified_multioutput}
\title{Prints model_unified_multioutput objects}
\usage{
\method{print}{model_unified_multioutput}(x, ...)
}
\arguments{
\item{x}{a model_unified_multioutput object}

\item{...}{other arguments}
}
\value{
No return value, called for printing
}
\description{
Prints model_unified_multioutput objects
}
