% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeshap.R
\name{print.treeshap}
\alias{print.treeshap}
\title{Prints treeshap objects}
\usage{
\method{print}{treeshap}(x, ...)
}
\arguments{
\item{x}{a model_unified object}

\item{...}{other arguments}
}
\description{
Prints treeshap objects
}
