% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeshap.R
\name{print.treeshap_multioutput}
\alias{print.treeshap_multioutput}
\title{Prints treeshap_multioutput objects}
\usage{
\method{print}{treeshap_multioutput}(x, ...)
}
\arguments{
\item{x}{a treeshap_multioutput object}

\item{...}{other arguments}
}
\value{
No return value, called for printing
}
\description{
Prints treeshap_multioutput objects
}
