% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_randomForest.R
\name{randomForest.unify}
\alias{randomForest.unify}
\title{Unify randomForest model}
\usage{
randomForest.unify(rf_model, data)
}
\arguments{
\item{rf_model}{An object of \code{randomForest} class. At the moment, models built on data with categorical features
are not supported - please encode them before training.}

\item{data}{data.frame for which calculations should be performed.}
}
\value{
a unified model representation - a \code{\link{model_unified.object}} object
}
\description{
Convert your randomForest model into a standarised data frame.
The returned data frame is easy to be interpreted by user and ready to be used as an argument in \code{treeshap()} function
}
\examples{

library(randomForest)
data_fifa <- fifa20$data[!colnames(fifa20$data) \%in\%
                           c('work_rate', 'value_eur', 'gk_diving', 'gk_handling',
                             'gk_kicking', 'gk_reflexes', 'gk_speed', 'gk_positioning')]
data <- na.omit(cbind(data_fifa, target = fifa20$target))

rf <- randomForest::randomForest(target~., data = data, maxnodes = 10, ntree = 10)
unified_model <- randomForest.unify(rf, data)
shaps <- treeshap(unified_model, data[1:2,])
# plot_contribution(shaps, obs = 1)

}
\seealso{
\code{\link{lightgbm.unify}} for \code{\code{\link[lightgbm:lightgbm]{LightGBM models}}}

\code{\link{gbm.unify}} for \code{\code{\link[gbm:gbm]{GBM models}}}

\code{\link{catboost.unify}} for  \code{\code{\link[catboost:catboost.train]{Catboost models}}}

\code{\link{xgboost.unify}} for \code{\code{\link[xgboost:xgboost]{XGBoost models}}}

\code{\link{ranger.unify}} for \code{\code{\link[ranger:ranger]{ranger models}}}
}
