% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_ranger_surv.R
\name{ranger_surv.unify}
\alias{ranger_surv.unify}
\title{Unify ranger survival model}
\usage{
ranger_surv.unify(rf_model, data, type = c("risk", "survival"))
}
\arguments{
\item{rf_model}{An object of \code{ranger} class. At the moment, models built on data with categorical features
are not supported - please encode them before training.}

\item{data}{Reference dataset. A \code{data.frame} or \code{matrix} with the same columns as in the training set of the model. Usually dataset used to train model.}

\item{type}{A character to define the type of prediction. Either \code{"risk"} (default),
which returns the cumulative hazards for each observation as risk score, or
\code{"survival"}, which predicts the survival probability at certain time-points for each observation.}
}
\value{
For \code{type = "risk"} a unified model representation is returned - a \code{\link{model_unified.object}} object.
For \code{type = "survival"} a list is returned that contains unified model representation ,
(\code{\link{model_unified.object}} objects) for each time point. In this case, the list names are the
\code{unique.death.times} (from the \code{ranger} object), at which the survival function was evaluated.
}
\description{
Convert your ranger model into a standardized representation.
The returned representation is easy to be interpreted by the user and ready to be used as an argument in \code{treeshap()} function.
}
\examples{

library(ranger)
data_colon <- data.table::data.table(survival::colon)
data_colon <- na.omit(data_colon[get("etype") == 2, ])
surv_cols <- c("status", "time", "rx")

feature_cols <- colnames(data_colon)[3:(ncol(data_colon) - 1)]

train_x <- model.matrix(
  ~ -1 + .,
  data_colon[, .SD, .SDcols = setdiff(feature_cols, surv_cols[1:2])]
)
train_y <- survival::Surv(
  event = (data_colon[, get("status")] |>
             as.character() |>
             as.integer()),
  time = data_colon[, get("time")],
  type = "right"
)

rf <- ranger::ranger(
  x = train_x,
  y = train_y,
  data = data_colon,
  max.depth = 10,
  num.trees = 10
)
unified_model_risk <- ranger_surv.unify(rf, train_x, type = "risk")
shaps <- treeshap(unified_model_risk, train_x[1:2,])


unified_model_surv <- ranger_surv_fun.unify(rf, train_x, type = "survival")
# compute shaps for first 3 death times
for (m in unified_model_surv[1:3]) {
  shaps <- treeshap(m, train_x[1:2,])
}

}
\seealso{
\code{\link{lightgbm.unify}} for \code{\link[lightgbm:lightgbm]{LightGBM models}}

\code{\link{gbm.unify}} for \code{\link[gbm:gbm]{GBM models}}

\code{\link{catboost.unify}} for \code{\link[catboost:catboost.train]{Catboost models}}

\code{\link{xgboost.unify}} for \code{\link[xgboost:xgboost]{XGBoost models}}

\code{\link{randomForest.unify}} for \code{\link[randomForest:randomForest]{randomForest models}}
}
