% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_covers.R
\name{recalculate_covers}
\alias{recalculate_covers}
\title{Compute Cover values of your model for reference dataset other than the original set used to train model.}
\usage{
recalculate_covers(model, x)
}
\arguments{
\item{model}{Unified dataframe representation of the model created with a (model).unify function.}

\item{x}{Reference dataset. A dataframe with the same columns as in the training set of the model.}
}
\value{
Unified dataframe representation of the model as created with a (model).unify function,
but with Cover column containing updated values.
}
\description{
Compute Cover values of your model for reference dataset other than the original set used to train model.
}
\examples{
\dontrun{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
             formula = value_eur ~ .,
             data = data,
             distribution = "laplace",
             n.trees = 1000,
             cv.folds = 2,
             interaction.depth = 2,
             n.cores = 1)
unified <- gbm.unify(gbm_model)
recalculate_covers(unified, data[200:700, ])
}
}
\seealso{
\code{\link{xgboost.unify}} for \code{XGBoost models}

\code{\link{lightgbm.unify}} for \code{LightGBM models}

\code{\link{catboost.unify}} for \code{Catboost models}
}
