% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeshap.R
\name{treeshap.object}
\alias{treeshap.object}
\title{treeshap results}
\value{
List consisting of four elements:
\describe{
\item{shaps}{A \code{data.frame} with M columns, X rows (M - number of features, X - number of explained observations). Every row corresponds to SHAP values for a observation. }
\item{interactions}{An \code{array} with dimesions (M, M, X) (M - number of features, X - number of explained observations). Every \code{[, , i]} slice is a symetric matrix - SHAP Interaction values for a observation. \code{[a, b, i]} element is SHAP Interaction value of features \code{a} and \code{b} for observation \code{i}. Is \code{NULL} if interactions where not calculated (parameter \code{interactions} set \code{FALSE}.) }
\item{unified_model}{An object of type \code{\link{model_unified.object}}. Unified representation of a model for which SHAP values were calculated. It is used by some of the plotting functions.}
\item{observations}{Explained dataset. \code{data.frame} or \code{matrix}. It is used by some of the plotting functions.}
}
}
\description{
\code{treeshap} object produced by \code{treeshap} function.
}
\seealso{
\code{\link{treeshap}}},

\code{\link{plot_contribution}}}, \code{\link{plot_feature_importance}}}, \code{\link{plot_feature_dependence}}}, \code{\link{plot_interaction}}}
}
