% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeshap.R
\name{treeshap_multioutput.object}
\alias{treeshap_multioutput.object}
\title{treeshap results for multi-output model}
\value{
List consisting of \code{treeshap} objects, one for each individual output of a model. For survival models, the list is named using the time points, for which TreeSHAP values are calculated.
}
\description{
\code{treeshap_multioutput} object produced by \code{treeshap} function.
}
\seealso{
\code{\link{treeshap}},

\code{\link{treeshap.object}}
}
