% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_catboost.R
\name{catboost.unify}
\alias{catboost.unify}
\title{Unify Catboost model}
\usage{
catboost.unify(catboost_model, data, recalculate = FALSE)
}
\arguments{
\item{catboost_model}{An object of \code{catboost.Model} class. At the moment, models built on data with categorical features
are not supported - please encode them before training.}

\item{data}{Reference dataset. A \code{data.frame} or \code{matrix} with the same columns as in the training set of the model. Usually dataset used to train model. Note that the same order of columns is crucial for unifier to work.}

\item{recalculate}{logical indicating if covers should be recalculated according to the dataset given in data. Keep it \code{FALSE} if training data is used.}
}
\value{
a unified model representation - a \code{\link{model_unified.object}} object
}
\description{
Convert your Catboost model into a standarised representation.
The returned representation is easy to be interpreted by the user and ready to be used as an argument in \code{treeshap()} function.
}
\examples{
if(requireNamespace("catboost")){
library(catboost)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data <- as.data.frame(lapply(data, as.numeric))
label <- fifa20$target
dt.pool <- catboost::catboost.load_pool(data = data, label = label)
cat_model <- catboost::catboost.train(
  dt.pool,
  params = list(loss_function = 'RMSE',
                iterations = 100,
                logging_level = 'Silent'))
um <- catboost.unify(cat_model, data)
shaps <- treeshap(um, data[1:2, ])
plot_contribution(shaps, obs = 1)
}
}
\seealso{
\code{\link{lightgbm.unify}} for \code{\link[lightgbm:lightgbm]{LightGBM models}}

\code{\link{gbm.unify}} for \code{\link[gbm:gbm]{GBM models}}

\code{\link{xgboost.unify}} for \code{\link[xgboost:xgboost]{XGBoost models}}

\code{\link{ranger.unify}} for \code{\link[ranger:ranger]{ranger models}}

\code{\link{randomForest.unify}} for \code{\link[randomForest:randomForest]{randomForest models}}
}
