% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifiers.R
\name{gbm.unify}
\alias{gbm.unify}
\title{Unify GBM model}
\usage{
gbm.unify(gbm_model, data)
}
\arguments{
\item{gbm_model}{An object of \code{gbm} class. At the moment, models built on data with categorical features
are not supported - please encode them before training.}

\item{data}{A training frame used to fit the model.}
}
\value{
Each row of a returned data frame indicates a specific node. The object has a defined structure:
\describe{
\item{Tree}{0-indexed ID of a tree}
\item{Node}{0-indexed ID of a node in a tree}
\item{Feature}{In case of an internal node - name of a feature to split on. Otherwise - NA}
\item{Split}{Threshold used for splitting observations.
All observations with lower or equal value than it are proceeded to the node marked as 'Yes'. Othwerwise to the 'No' node}
\item{Yes}{Index of a row containing a child Node. Thanks to explicit indicating the row it is much faster to move between nodes.}
\item{No}{Index of a row containing a child Node}
\item{Missing}{Index of a row containing a child Node where are proceeded all observations with no value of the dividing feature}
\item{Quality/Score}{For internal nodes - Quality: the reduction in the loss function as a result of splitting this node.
For leaves - Score: Value of prediction in the leaf}
\item{Cover}{Number of observations seen by the internal node or collected by the leaf}
}
}
\description{
Convert your GBM model into a standarised data frame.
The returned data frame is easy to be interpreted by user and ready to be used as an argument in \code{treeshap()} function
}
\examples{
\dontrun{
library(gbm)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
data['value_eur'] <- fifa20$target
gbm_model <- gbm::gbm(
             formula = value_eur ~ .,
             data = data,
             distribution = "laplace",
             n.trees = 1000,
             cv.folds = 2,
             interaction.depth = 2,
             n.cores = 1)
gbm.unify(gbm_model)
}
}
\seealso{
\code{\link{xgboost.unify}} for \code{XGBoost models}

\code{\link{lightgbm.unify}} for \code{LightGBM models}

\code{\link{catboost.unify}} for \code{Catboost models}
}
