% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_ranger.R
\name{ranger.unify}
\alias{ranger.unify}
\title{Unify ranger model}
\usage{
ranger.unify(rf_model, data)
}
\arguments{
\item{rf_model}{An object of \code{ranger} class. At the moment, models built on data with categorical features
are not supported - please encode them before training.}

\item{data}{data.frame for which calculations should be performed.}
}
\value{
Each row of a returned data frame indicates a specific node. The object has a defined structure:
\describe{
\item{Tree}{0-indexed ID of a tree}
\item{Node}{0-indexed ID of a node in a tree}
\item{Feature}{In case of an internal node - name of a feature to split on. Otherwise - NA}
\item{Split}{Threshold used for splitting observations.
All observations with lower or equal value than it are proceeded to the node marked as 'Yes'. Othwerwise to the 'No' node}
\item{Yes}{Index of a row containing a child Node. Thanks to explicit indicating the row it is much faster to move between nodes.}
\item{No}{Index of a row containing a child Node}
\item{Missing}{Index of a row containing a child Node where are proceeded all observations with no value of the dividing feature}
\item{Prediction}{For leaves: Value of prediction in the leaf. For internal nodes: NA.}
\item{Cover}{Number of observations seen by the internal node or collected by the leaf}
}
}
\description{
Convert your ranger model into a standarised data frame.
The returned data frame is easy to be interpreted by user and ready to be used as an argument in \code{treeshap()} function
}
\examples{

 library(ranger)
 data_fifa <- fifa20$data[!colnames(fifa20$data) \%in\%
                            c('work_rate', 'value_eur', 'gk_diving', 'gk_handling',
                             'gk_kicking', 'gk_reflexes', 'gk_speed', 'gk_positioning')]
 data <- na.omit(cbind(data_fifa, target = fifa20$target))

 rf <- ranger::ranger(target~., data = data, max.depth = 10, num.trees = 10)
 unified_model <- ranger.unify(rf, data)
 shaps <- treeshap(unified_model, data[1:2,])
 plot_contribution(shaps, obs = 1)
}
\seealso{
\code{\link{xgboost.unify}} for \code{XGBoost models}

\code{\link{lightgbm.unify}} for \code{LightGBM models}

\code{\link{catboost.unify}} for \code{Catboost models}
}
