% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeshap.R
\name{treeshap}
\alias{treeshap}
\title{Calculate SHAP values of a tree ensemble model.}
\usage{
treeshap(unified_model, x, interactions = FALSE, verbose = TRUE)
}
\arguments{
\item{unified_model}{Unified data.frame representation of the model created with a (model).unify function.}

\item{x}{Observations to be explained. A data.frame object with the same columns as in the training set of the model. Keep in mind that objects different than data.frame or plain matrix will cause an error or unpredictable behaviour.}

\item{interactions}{Whether to calculate SHAP interaction values. By default is \code{FALSE}.}

\item{verbose}{Wheter to print progress bar to the console.}
}
\value{
A treeshap object
}
\description{
Calculate SHAP values and optionally SHAP Interaction values.
}
\examples{
\dontrun{
library(xgboost)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
target <- fifa20$target

# calculating simple SHAP values
param <- list(objective = "reg:squarederror", max_depth = 3)
xgb_model <- xgboost::xgboost(as.matrix(data), params = param, label = target, nrounds = 200,
                              verbose = 0)
unified_model <- xgboost.unify(xgb_model, as.matrix(data))
shaps <- treeshap(unified_model, head(data, 3))
plot_contribution(shaps, obs = 1)

# It's possible to calcualte explanation over different part of the data set

unified_model_rec <- recalculate_covers(unified_model, data[1:1000, ])
shaps_rec <- treeshap(unified_model, head(data, 3))
plot_contribution(shaps_rec, obs = 1)

# calculating SHAP interaction values
param2 <- list(objective = "reg:squarederror", max_depth = 20)
xgb_model2 <- xgboost::xgboost(as.matrix(data), params = param2, label = target, nrounds = 10)
unified_model2 <- xgboost.unify(xgb_model2, as.matrix(data))
treeshap(unified_model2, head(data, 3), interactions = TRUE)
}
}
\seealso{
\code{\link{xgboost.unify}} for \code{XGBoost models}
\code{\link{lightgbm.unify}} for \code{LightGBM models}
\code{\link{gbm.unify}} for \code{GBM models}
\code{\link{catboost.unify}} for \code{catboost models}
\code{\link{randomForest.unify}} for \code{randomForest models}
\code{\link{ranger.unify}} for \code{ranger models}
}
