% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_xgboost.R
\name{xgboost.unify}
\alias{xgboost.unify}
\title{Unify xgboost model}
\usage{
xgboost.unify(xgb_model, data, recalculate = FALSE)
}
\arguments{
\item{xgb_model}{A xgboost model - object of class \code{xgb.Booster}}

\item{data}{matrix for which calculations should be performed.}

\item{recalculate}{logical indicating if covers should be recalculated according to the dataset given in data. Keep it FALSE if training data are used.}
}
\value{
Each row of a returned data frame indicates a specific node. The object has a defined structure:
\describe{
\item{Tree}{0-indexed ID of a tree}
\item{Node}{0-indexed ID of a node in a tree}
\item{Feature}{In case of an internal node - name of a feature to split on. Otherwise - NA.}
\item{Split}{Threshold used for splitting observations.
All observations with lower or equal value than it are proceeded to the node marked as 'Yes'. Otherwise to the 'No' node}
\item{Yes}{Index of a row containing a child Node. Thanks to explicit indicating the row it is much faster to move between nodes.}
\item{No}{Index of a row containing a child Node}
\item{Missing}{Index of a row containing a child Node where are proceeded all observations with no value of the dividing feature}
\item{Prediction}{For leaves: Value of prediction in the leaf. For internal nodes: NA.}
\item{Cover}{Number of observations seen by the internal node or collected by the leaf}
}
}
\description{
Convert your xgboost model into a standarised data frame.
The returned data frame is easy to be interpreted by user and ready to be used as an argument in the \code{treeshap()} function.
}
\examples{
library(xgboost)
data <- fifa20$data[colnames(fifa20$data) != 'work_rate']
target <- fifa20$target
param <- list(objective = "reg:squarederror", max_depth = 3)
xgb_model <- xgboost::xgboost(as.matrix(data), params = param, label = target,
                              nrounds = 200, verbose = 0)
unified_model <- xgboost.unify(xgb_model, as.matrix(data))
shaps <- treeshap(unified_model, data[1:2,])
plot_contribution(shaps, obs = 1)

}
\seealso{
\code{\link{lightgbm.unify}} for \code{LightGBM models}

\code{\link{gbm.unify}} for \code{GBM models}

\code{\link{catboost.unify}} for \code{Catboost models}
}
